requirements-dev.txt
# Development dependencies
pytest==8.3.5
pytest-cov==6.0.0
pytest-playwright==0.6.2
flake8==7.0.0
black==24.1.1
